% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.R
\name{.trimNA}
\alias{.trimNA}
\title{remove NA values at the start and end}
\usage{
.trimNA(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
subset of x with leading and trailing NAs removed
}
\description{
remove NA values at the start and end
}
\keyword{internal}
