% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormalDiff.R
\name{estimateDiffLognormal}
\alias{estimateDiffLognormal}
\title{Estimatd the shifted-lognormal approximation to difference of two lognormals}
\usage{
estimateDiffLognormal(mu_a, mu_b, sigma_a, sigma_b, corr = 0)
}
\arguments{
\item{mu_a}{center parameter of the first term}

\item{mu_b}{center parameter of the second term}

\item{sigma_a}{scale parameter of the first term}

\item{sigma_b}{scale parameter of the second term}

\item{corr}{correlation between the two random variables}
}
\value{
numeric vector with components mu, sigma, and shift, the components
of the shifted lognormal distribution.
}
\description{
The distribtuion of y = a - b + s, where a and b are two lognormal random
variables and s is a constant to be estimated, can be approximated
by a lognormal distribution.
}
