% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefLognorm.R
\name{estimateParmsLognormFromSample}
\alias{estimateParmsLognormFromSample}
\title{Estimate lognormal distribution parameters from a sample}
\usage{
estimateParmsLognormFromSample(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of sampled values}

\item{na.rm}{a logical value indicating whether 
NA values should be stripped before the computation proceeds.}
}
\value{
numeric vector with components \code{mu} and \code{sigma},
ie.., the center parameter (mean at log scale, log(median)) and 
the scale parameter (standard deviation at log scale)
}
\description{
Estimate lognormal distribution parameters from a sample
}
