% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.R
\name{getCorrMatFromAcf}
\alias{getCorrMatFromAcf}
\title{Construct the full correlation matrix from autocorrelation components.}
\usage{
getCorrMatFromAcf(nRow, effAcf)
}
\arguments{
\item{nRow}{number of rows in correlation matrix}

\item{effAcf}{numeric vector of effective autocorrelation components
. The first entry, which is defined as 1, is not used.}
}
\description{
Construct the full correlation matrix from autocorrelation components.
}
