\name{getLognormMedian}
\alias{getLognormMedian}
\title{getLognormMedian}
\description{get the median of a log-normal distribution}
\usage{getLognormMedian(mu, sigma = NA)}
\arguments{
  \item{mu}{center parameter (mean at log scale, log(median))}
  \item{sigma}{dummy not used, but signature as with Mode and moments}
}

\value{the median}

\author{Thomas Wutzler}





