\name{getParmsLognormForExpval}
\alias{getParmsLognormForExpval}
\title{getParmsLognormForExpval}
\description{get the lognormal parameters by expected value}
\usage{getParmsLognormForExpval(mean, sigmaStar)}
\arguments{
  \item{mean}{expected value at original scale}
  \item{sigmaStar}{multiplicative standard deviation}
}



\author{Thomas Wutzler}





