% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.R
\name{computeEffectiveNumObs}
\alias{computeEffectiveNumObs}
\title{Compute the effective number of observations taking into account autocorrelation}
\usage{
computeEffectiveNumObs(
  res,
  effAcf = computeEffectiveAutoCorr(res),
  na.rm = FALSE
)
}
\arguments{
\item{res}{numeric of autocorrelated numbers, usually observation -
model residuals}

\item{effAcf}{autocorrelation coefficients.
The first entry is fixed at 1 for zero distance.}

\item{na.rm}{a logical value indicating whether NA values should be 
stripped before the computation proceeds.}
}
\value{
integer scalar: effective number of observations
}
\description{
Compute the effective number of observations taking into account autocorrelation
}
\details{
Handling of NA values: NAs at the beginning or end are 
just trimmed before computation and pose no problem. 
However with NAs aside from edges, the return value is biased low,
because correlation terms are subtracted for those positions.

Because of NA correlation terms, the computed effective number of
observations can be smaller than 1. In this case 1 is returned.
}
\references{
\code{Zieba & Ramza (2011) 
Standard Deviation of the Mean of Autocorrelated 
Observations Estimated with the Use of the Autocorrelation Function 
Estimated From the Data. 
Metrology and Measurement Systems, 
Walter de Gruyter GmbH, 18 10.2478/v10178-011-0052-x}
\code{Bayley & Hammersley (1946) 
The "effective" number of independent observations in an autocorrelated 
time series. 
Supplement to the Journal of the Royal Statistical Society, JSTOR,8,184-197}
}
