\name{estimateParmsLognormFromSample}
\alias{estimateParmsLognormFromSample}
\title{estimateParmsLognormFromSample}
\description{get the lognormal parameters by expected value.}
\usage{estimateParmsLognormFromSample(x, na.rm = FALSE)}
\arguments{
  \item{x}{numeric vector of sampled values}
  \item{na.rm}{a logical value indicating whether
NA values should be stripped before the computation proceeds.}
}



\author{Thomas Wutzler}





