\name{getCorrMatFromAcf}
\alias{getCorrMatFromAcf}
\title{getCorrMatFromAcf}
\description{Construct the full correlation matrix from autocorrelation components.}
\usage{getCorrMatFromAcf(nRow, effAcf)}
\arguments{
  \item{nRow}{number of rows in correlation matrix}
  \item{effAcf}{numeric vector of effective autocorrelation components}
}



\author{Thomas Wutzler}





