\name{getLognormMode}
\alias{getLognormMode}
\title{getLognormMode}
\description{get the mode of a log-normal distribution}
\usage{getLognormMode(mu, sigma)}
\arguments{
  \item{mu}{center parameter (mean at log scale, log(median))}
  \item{sigma}{scale parameter (standard deviation at log scale)}
}

\value{the mode}

\author{Thomas Wutzler}





