% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.R
\name{seCor}
\alias{seCor}
\title{Compute the standard error accounting for empirical autocorrelations}
\usage{
seCor(
  x,
  na.rm = FALSE,
  effCov = computeEffectiveAutoCorr(x, type = "covariance")
)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical. Should missing values be removed?}

\item{effCov}{numeric vector of effective covariance components
first entry is the variance. See \code{\link{computeEffectiveAutoCorr}}}
}
\value{
numeric scalar of standard error of the mean of x
}
\description{
Compute the standard error accounting for empirical autocorrelations
}
\details{
Computation follows 
 https://stats.stackexchange.com/questions/274635/calculating-error-of-mean-of-time-series.

The default uses empirical autocorrelation
 estimates from the supplied data up to first negative component.
 For short series of \code{x} it is strongly recommended to to
 provide \code{effCov} that was estimated on a longer time series.
}
